Allow filename override
Page objects can have a filename: key to override the (base) name of the
image file(s). Useful for authors that prefer to name their files along
patterns such as "chapter##_page##" etc rather than just sequentially
numbered across all pages.

This is not documented in the example settings, though, and I'm not sure
what's the best way to do it since it's optional.